+++ rpmsetcmp.c
@@ -145,14 +145,14 @@ static inline void install_sentinels(unsigned *v, int n)
  * It suffices to merge v1[] and v2[].
  */
 /* Reduce a set of (bpp + 1) values to a set of bpp values. */
-static int downsample1(const unsigned *v, int n, unsigned *w, int bpp)
+static int downsample1(const unsigned *v, size_t n, unsigned *w, int bpp)
 {
     unsigned mask = (1U << bpp) - 1;
     /* Find the first element with high bit set. */
-    int l = 0;
-    int u = n;
+    size_t l = 0;
+    size_t u = n;
     while (l < u) {
-	int i = (l + u) / 2;
+	size_t i = (l + u) / 2;
 	if (v[i] <= mask)
 	    l = i + 1;
 	else
